+++ threedi_api_client/openapi/models/threedi_model.py
@@ -3,7 +3,7 @@
 """
     3Di API
 
-    3Di simulation API (latest stable version: v3)   Framework release: 2.1.1   3Di core release: 2.1.9  deployed on:  10:36AM (UTC) on December 09, 2021  # noqa: E501
+    3Di simulation API (latest stable version: v3)   Framework release: 2.7.0   3Di core release: 2.2.2  deployed on:  01:00PM (UTC) on January 10, 2022  # noqa: E501
 
     The version of the OpenAPI document: v3
     Contact: info@nelen-schuurmans.nl
@@ -479,8 +479,8 @@ class ThreediModel(object):
         if self.local_vars_configuration.client_side_validation and slug is None:  # noqa: E501
             raise ValueError("Invalid value for `slug`, must not be `None`")  # noqa: E501
         if (self.local_vars_configuration.client_side_validation and
-                slug is not None and len(slug) > 255):
-            raise ValueError("Invalid value for `slug`, length must be less than or equal to `255`")  # noqa: E501
+                slug is not None and len(slug) > 512):
+            raise ValueError("Invalid value for `slug`, length must be less than or equal to `512`")  # noqa: E501
         if (self.local_vars_configuration.client_side_validation and
                 slug is not None and len(slug) < 1):
             raise ValueError("Invalid value for `slug`, length must be greater than or equal to `1`")  # noqa: E501